requirements.txt
streamlit
requests
libsql-client
python-dotenv
